Enabling more fail on error scenarios (#4286)
Turning on a number of compiler linting errors so we can be more stringent. Suppressing a small number of the failures, and fixing the rest.

+++ core/azure-core/src/main/java/com/azure/core/implementation/util/FluxUtil.java
@@ -371,7 +371,7 @@ public final class FluxUtil {
                         // use local variable to perform fewer volatile reads
                         long pos = position;
                         //
-                        int bytesWanted = (int) Math.min(bytesRead, maxRequired(pos));
+                        int bytesWanted = Math.min(bytesRead, maxRequired(pos));
                         buffer.writerIndex(bytesWanted);
                         long position2 = pos + bytesWanted;
                         //noinspection NonAtomicOperationOnVolatileField
